batch-prediction-pipeline/pyproject.toml
[tool.poetry]
name = "batch-prediction-pipeline"
version = "0.1.0"
description = ""
authors = ["Iusztin Paul <p.e.iusztin@gmail.com>"]
readme = "README.md"
packages = [{include = "batch_prediction_pipeline"}]

[tool.poetry.dependencies]
python = "~3.11"
category-encoders = "^2.6.0"
hopsworks = "3.4.3"
python-dotenv = "^1.0.0"
lightgbm = "^3.3.5"
sktime = "^0.16.1"
google-cloud-storage = "^2.7.0"
fire = "^0.5.0"
s3fs = {extras = ["boto3"], version = ">=2023.12.0"}
boto3 = "*"
botocore = "*"
training-pipeline = "^0.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipykernel = "^6.20.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "my-pypi"
url = "http://192.168.192.27:8090"
priority = "primary"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

